webapp-requirements.txt
fastapi
uvicorn[standard]
pyyaml
python-multipart
salsa20
cantools
numpy
pandas